requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9

# AI & NLP
openai==1.30.1
transformers==4.41.1
torch==2.3.0
sentence-transformers==3.0.0
tokenizers==0.19.1

# Vector Search
faiss-cpu==1.8.0
numpy==1.26.4

# AST Parsing
tree-sitter==0.22.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.4
tree-sitter-java==0.21.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Background Jobs
celery==5.4.0
redis==5.0.4
flower==2.0.1

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5

# Code Quality
radon==6.0.1
pylint==3.2.2
bandit==1.7.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
rich==13.7.1
typer==0.12.3
networkx==3.3
matplotlib==3.9.0
scipy==1.13.0

# Serialization
orjson==3.10.3
